refactor: extract SeriesBuffer into module
